First result
My task was about to analyze data in Online_retail table and create any type of presentation to present my results. I used PowerPoint for presentation and PowerBI as tool, which help me to do some basic graphs for data, which were better to show in visualization, which are possible to see in this file.